invoice-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
pdf2image
pdfplumber
lxml